粒度を下げた
