+++ front/app/layout.tsx
@@ -4,14 +4,13 @@ import type { Metadata } from "next"
 import { Inter } from "next/font/google"
 import {
   ClerkProvider,
-  SignInButton,
-  SignUpButton,
   SignedIn,
   SignedOut,
-  UserButton
 } from "@clerk/nextjs"
 import { ThemeProvider } from "@/components/theme-provider"
 import { Toaster } from "@/components/ui/toaster"
+import { AuthenticatedLayout } from "@/components/layout/authenticated-layout"
+import { Header } from "@/components/layout/header"
 
 const inter = Inter({ subsets: ["latin"] })
 
@@ -31,35 +30,13 @@ export default function RootLayout({
       <html lang="ja">
         <body className={inter.className}>
           <ThemeProvider attribute="class" defaultTheme="light" enableSystem disableTransitionOnChange>
-            <header className="border-b px-4 py-3">
-              <div className="max-w-7xl mx-auto flex justify-between items-center">
-                <h1 className="text-xl font-semibold">就活ダッシュボード</h1>
-                <div className="flex items-center gap-4">
-                  <SignedOut>
-                    <SignInButton mode="modal">
-                      <button className="bg-blue-600 hover:bg-blue-700 text-white px-4 py-2 rounded-md transition-colors">
-                        サインイン
-                      </button>
-                    </SignInButton>
-                    <SignUpButton mode="modal">
-                      <button className="border border-gray-300 hover:bg-gray-50 px-4 py-2 rounded-md transition-colors">
-                        新規登録
-                      </button>
-                    </SignUpButton>
-                  </SignedOut>
-                  <SignedIn>
-                    <UserButton
-                      appearance={{
-                        elements: {
-                          avatarBox: "w-8 h-8"
-                        }
-                      }}
-                    />
-                  </SignedIn>
-                </div>
-              </div>
-            </header>
-            {children}
+            <Header />
+            <SignedOut>
+              {children}
+            </SignedOut>
+            <SignedIn>
+              <AuthenticatedLayout>{children}</AuthenticatedLayout>
+            </SignedIn>
             <Toaster />
           </ThemeProvider>
         </body>
