+++ front/components/es/es-manager.tsx
