+++ front/components/layout/authenticated-layout.tsx
@@ -1,0 +1,16 @@
+"use client"
+
+import { Sidebar } from "@/components/layout/sidebar"
+
+interface AuthenticatedLayoutProps {
+  children: React.ReactNode
+}
+
+export function AuthenticatedLayout({ children }: AuthenticatedLayoutProps) {
+  return (
+    <div className="flex h-screen bg-gray-50">
+      <Sidebar />
+      {children}
+    </div>
+  )
+}
