+++ front/components/layout/header.tsx
@@ -1,0 +1,34 @@
+import { SignedIn, SignedOut, SignInButton, SignUpButton, UserButton } from "@clerk/nextjs"
+
+export function Header() {
+  return (
+    <header className="border-b px-4 py-3">
+      <div className="max-w-7xl mx-auto flex justify-between items-center">
+        <h1 className="text-xl font-semibold">就活サポーター</h1>
+        <div className="flex items-center gap-4">
+          <SignedOut>
+            <SignInButton mode="modal">
+              <button className="bg-blue-600 hover:bg-blue-700 text-white px-4 py-2 rounded-md transition-colors">
+                サインイン
+              </button>
+            </SignInButton>
+            <SignUpButton mode="modal">
+              <button className="border border-gray-300 hover:bg-gray-50 px-4 py-2 rounded-md transition-colors">
+                新規登録
+              </button>
+            </SignUpButton>
+          </SignedOut>
+          <SignedIn>
+            <UserButton
+              appearance={{
+                elements: {
+                  avatarBox: "w-8 h-8"
+                }
+              }}
+            />
+          </SignedIn>
+        </div>
+      </div>
+    </header>
+  )
+}
